+++ AOS/app/src/main/java/com/softeer/team6four/ui/mypage/charger/MyChargerDetailViewModel.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.softeer.team6four.data.Resource
-import com.softeer.team6four.data.local.UserPreferencesRepository
-import com.softeer.team6four.data.remote.charger.ChargerRepository
+import com.softeer.team6four.data.UserPreferencesRepository
+import com.softeer.team6four.data.ChargerRepository
 import com.softeer.team6four.data.remote.charger.model.MyChargerDetailInfoModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
