+++ AOS/app/src/main/java/com/softeer/team6four/ui/user/signup/SignUpViewModel.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.softeer.team6four.data.Resource
-import com.softeer.team6four.data.remote.user.UserRepository
+import com.softeer.team6four.data.UserRepository
 import com.softeer.team6four.data.remote.user.model.UserInfoModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
